:recycle: [#87] component 구조 변경
