+++ trecord/src/pages/FeedDetail/index.tsx
@@ -11,7 +11,7 @@ import { ReactElement, useCallback, useMemo, useState } from 'react';
 import Modal from '@components/common/Modal';
 import { SELECT_INFOS } from '@/types';
 import Skeleton from '@components/common/skeleton';
-import ShareModalBody from '@components/common/Modal/ShareModalBody';
+import ShareModalBody from '@components/common/Modal/ModalBody/ShareModalBody';
 import useGetRecordList from '@/apis/Feed/getRecordList';
 import ChipContainer from '@components/common/ChipContainer';
